+++ Controllers/CryorigH7QuadLumiController/RGBController_CryorigH7QuadLumi.cpp
@@ -240,7 +240,7 @@ void RGBController_CryorigH7QuadLumi::DeviceUpdateMode()
                     modes[active_mode].speed,
                     direction,
                     colors,
-                    modes[active_mode].colors.size()
+                    (unsigned int)modes[active_mode].colors.size()
                     );
         }
     }
